requirements.txt
# FastAPI and dependencies
fastapi>=0.95.2
uvicorn[standard]>=0.27.1
python-multipart>=0.0.5
email-validator>=2.1.0.post1

# Database and ORM
psycopg>=3.2.3
psycopg-binary>=3.2.3
SQLAlchemy>=2.0.27
alembic>=1.13.1

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.0.0
python-decouple>=3.8

# Authentication and security
bcrypt>=4.2.1
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Task scheduling
schedule>=1.2.1

# Template engine
Jinja2>=3.1.3

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.1
httpx>=0.24.1
coverage>=7.6.8
